% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dciu_rbgl.R
\name{dciu_rbgl}
\alias{dciu_rbgl}
\title{Directional connectivity index distance looping}
\usage{
dciu_rbgl(start_nodes, adjacency, dist, pixelx, R, dx, adjL)
}
\arguments{
\item{start_nodes}{list of nodes}

\item{adjacency}{matrix representing adjacency among nodes}

\item{dist}{a distance matrix taking into account longer diagonal distances}

\item{pixelx}{the rows positions of each node}

\item{R}{integer number of matrix ros}

\item{dx}{numeric cell (grain) distance}

\item{adjL}{list of adjacent nodes}
}
\description{
Directional connectivity index distance looping
}

