% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dciu_rbgl.R
\name{dciu_rbgl}
\alias{dciu_rbgl}
\title{Directional connectivity index distance looping}
\usage{
dciu_rbgl(start_nodes, adjacency, dist, pixelx, R, dx, adjL)
}
\description{
Directional connectivity index distance looping
}
\examples{
mat <- matrix(c(0,0,1,0,0,0,1,0,1,0,0,1,0,0,1,0,0,0,1,0,1,1,0,0,0),nrow=5,byrow=T)
xgrain<-1
ygrain<-1
#run first part of dci.R through dciu_pre
adjacency<-adj$adjacency
start_nodes<-dciu_pre$start_nodes
pixelx<-adj$pixelx
R<-dciu_pre$R
dist<-adj$distance
dx<-dciu_pre$dx
}

